+++ jest.coverage.config.ts
@@ -2,6 +2,6 @@ import type {Config} from 'jest';
 
 import {createJestConfig} from './jest/create-jest-config';
 
-const config: Config = createJestConfig({collectCoverage: false});
+const config: Config = createJestConfig({collectCoverage: true});
 
 export default config;